Updated imports, exports and pause menu spacing
