+++ src/main.js
@@ -4,8 +4,7 @@ import { GameOver } from './scenes/GameOver.js';
 import { MainMenu } from './scenes/MainMenu.js';
 import { Preloader } from './scenes/Preloader.js';
 import { Credits } from './scenes/Credits.js';
-// import { PauseMenu } from './scenes/PauseMenu.js';
-import { PauseMenu } from './scenes/PauseMenuNext.js';
+import { PauseMenu } from './scenes/PauseMenu.js';
 import { HUD } from './scenes/HUD.js';
 import { AUTO, Scale, Game } from 'phaser';
 import { TOUCH_CONTROLS_KEY } from './constants/data.js';
